chore(discord): decompose turn_finalizer.rs under giant baseline (#3894)
Extract three cohesive clusters VERBATIM (pure move, zero logic change) out of
the #3016 core hotfile `src/services/discord/turn_finalizer.rs` into sibling
turn_finalizer/ child modules, bringing it back under the giant baseline that
#3866/#3887 raised to 1521 prod LoC for the panic-containment safety fix.

Clusters moved (each well under the 1000-prod-LoC giant threshold):
- turn_finalizer/finalize_context.rs (106 prod LoC): the per-submission
  `FinalizeContext` knobs struct + constructor set (bridge/watcher/monitor/
  gate_backstop/delivery_lease). Re-exported so call sites are unchanged.
- turn_finalizer/finalize.rs (208 prod LoC): `do_finalize`, the single finalize
  side-effect chokepoint. Re-imported so handle_terminal + the reconcile child
  call it byte-identically.
- turn_finalizer/reconcile.rs (220 prod LoC): the timer-driven `reconcile` pass
  + `run_backstop_finalize` helper. `reconcile` re-imported so the actor loop's
  select! arm is unchanged.

The only edits to the moved bodies are mechanical and necessary for the
relocation: re-export/re-import for unchanged call sites, widening
`FinalizeContext::gate_backstop` private -> pub(super), and rewriting bare
`super::` discord-module references to absolute `crate::services::discord::`
paths (house style) since the module depth changed. Reverse-transform diffs
confirm the bodies are otherwise byte-identical to the originals.

Baseline lowered in scripts/audit_maintainability_giant_baseline.toml:
turn_finalizer.rs 1521 -> 1038 (-483 prod LoC); the three children tracked at
landed LoC so they cannot silently re-inflate toward giant-hood.

Behavior-preserving proof: cargo test --lib turn_finalizer = 90 passed / 0
failed, including the #3866 panic-containment tests (handle_terminal,
do_finalize stuck-Finalizing reset, reconcile/backstop). Gates: cargo fmt
--check clean; cargo check --lib clean; check_agent_maintenance_docs passed;
check_hotfile_ratchet exit 0; audit_maintainability --check exit 0.

Note: turn_finalizer.rs is not tracked in scripts/hotfile_ratchet.toml (only
tmux_watcher.rs, tui_prompt_relay.rs, turn_bridge/mod.rs are), so there is no
raw-LOC ceiling to lower there; the production-LoC giant baseline is the
applicable ratchet and it is lowered.
