+++ ARCHITECTURE.md
@@ -654,6 +654,9 @@ src/
 │   │   │   ├── cleanup.rs
 │   │   │   ├── completion_signal.rs
 │   │   │   ├── delivery_lease.rs
+│   │   │   ├── finalize.rs
+│   │   │   ├── finalize_context.rs
+│   │   │   ├── reconcile.rs
 │   │   │   └── watcher_backstop.rs
 │   │   ├── voice_barge_in/
 │   │   │   ├── tests/
